+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Fixed:
 Changed:
 
   * :fire: `OcrdFile.local_filename` returns/accepts `str` after unpopular change to `Path` from #1079, #1182, #1167
+  * `WorkspaceValidator`: more efficiency by doing all page checks in the same loop, #1071
+
+Added:
+
+  * `OcrdMets.get_physical_pages` to search for/change/generate ranges for page-specific `mets:div` attributes beyond to `@ID`, #821, #1063
 
 ## [2.62.0] - 2024-01-30